+++ spec/spec_helper.rb
@@ -11,6 +11,7 @@ LATEST_PYTHON_3_7 = '3.7.15'
 LATEST_PYTHON_3_8 = '3.8.15'
 LATEST_PYTHON_3_9 = '3.9.15'
 LATEST_PYTHON_3_10 = '3.10.8'
+LATEST_PYTHON_3_11 = '3.11.0'
 DEFAULT_PYTHON_VERSION = LATEST_PYTHON_3_10
 
 # Work around the return value for `default_buildpack` changing after deploy: