+++ spec/spec_helper.rb
@@ -10,8 +10,8 @@ LATEST_PYTHON_3_6 = '3.6.15'
 LATEST_PYTHON_3_7 = '3.7.16'
 LATEST_PYTHON_3_8 = '3.8.16'
 LATEST_PYTHON_3_9 = '3.9.16'
-LATEST_PYTHON_3_10 = '3.10.9'
-LATEST_PYTHON_3_11 = '3.11.1'
+LATEST_PYTHON_3_10 = '3.10.10'
+LATEST_PYTHON_3_11 = '3.11.2'
 DEFAULT_PYTHON_VERSION = LATEST_PYTHON_3_10
 
 # Work around the return value for `default_buildpack` changing after deploy: